Добавить всякое разное, и генерацию proto через makefile

+++ appserver/views/userinfo.go
@@ -2,5 +2,5 @@ package views
 
 type UserInfo struct {
 	UserName string `json:"username"`
-	CountEvens int  `json:"count"`
+	CountEvens int64  `json:"count"`
 }
